Re-initialize git repo
This is done because I accidently pulled in a lot of repositories as
subtrees. Two out of three included repositories were not even relevant for
development. In order to keep the history small I'll squash the current
history into one commit and point to the Human Connection API repository
as a git submodule. A submodule is a better fit here, as we only want to
point to a commit, not copy the entire history. Pushing back changes
into the HC Api repository is not very likely to happen.